refactor(db): move sqlcomment under pkg/db/sqlcomment
