+++ pkg/db/traced_tx.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/unkeyed/unkey/pkg/mysql/metrics"
 	"github.com/unkeyed/unkey/pkg/otel/tracing"
-	"github.com/unkeyed/unkey/pkg/sqlcomment"
+	"github.com/unkeyed/unkey/pkg/db/sqlcomment"
 	"go.opentelemetry.io/otel/attribute"
 )
