+++ pkg/mysql/BUILD.bazel
@@ -19,12 +19,12 @@ go_library(
     deps = [
         "//pkg/assert",
         "//pkg/codes",
+        "//pkg/db/sqlcomment",
         "//pkg/fault",
         "//pkg/logger",
         "//pkg/mysql/metrics",
         "//pkg/otel/tracing",
         "//pkg/retry",
-        "//pkg/sqlcomment",
         "@com_github_go_sql_driver_mysql//:mysql",
         "@io_opentelemetry_go_otel//attribute",
     ],
