+++ svc/api/BUILD.bazel
@@ -34,6 +34,7 @@ go_library(
         "//pkg/config",
         "//pkg/counter",
         "//pkg/db",
+        "//pkg/db/sqlcomment",
         "//pkg/logger",
         "//pkg/otel",
         "//pkg/prometheus",
@@ -41,7 +42,6 @@ go_library(
         "//pkg/rbac",
         "//pkg/rpc/interceptor",
         "//pkg/runner",
-        "//pkg/sqlcomment",
         "//pkg/tls",
         "//pkg/uid",
         "//pkg/zen",
