+++ svc/ctrl/api/BUILD.bazel
@@ -22,6 +22,7 @@ go_library(
         "//pkg/clickhouse/schema",
         "//pkg/clock",
         "//pkg/config",
+        "//pkg/db/sqlcomment",
         "//pkg/dns/domainconnect",
         "//pkg/logger",
         "//pkg/otel",
@@ -29,7 +30,6 @@ go_library(
         "//pkg/prometheus/lazy",
         "//pkg/restate/admin",
         "//pkg/runner",
-        "//pkg/sqlcomment",
         "//pkg/uid",
         "//svc/ctrl/internal/auditlogs",
         "//svc/ctrl/internal/db",
@@ -66,9 +66,9 @@ go_test(
         "//gen/proto/ctrl/v1/ctrlv1connect",
         "//gen/proto/hydra/v1:hydra",
         "//pkg/config",
+        "//pkg/db/sqlcomment",
         "//pkg/logger",
         "//pkg/rpc/interceptor",
-        "//pkg/sqlcomment",
         "//pkg/testutil/containers",
         "//pkg/uid",
         "//svc/ctrl/integration/seed",
