+++ svc/ctrl/integration/BUILD.bazel
@@ -6,8 +6,8 @@ go_library(
     importpath = "github.com/unkeyed/unkey/svc/ctrl/integration",
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/db/sqlcomment",
         "//pkg/mysql/types",
-        "//pkg/sqlcomment",
         "//pkg/testutil/containers",
         "//pkg/uid",
         "//svc/ctrl/integration/seed",
