+++ svc/ctrl/integration/harness.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/stretchr/testify/require"
 	dbtype "github.com/unkeyed/unkey/pkg/mysql/types"
-	"github.com/unkeyed/unkey/pkg/sqlcomment"
+	"github.com/unkeyed/unkey/pkg/db/sqlcomment"
 	"github.com/unkeyed/unkey/pkg/testutil/containers"
 	"github.com/unkeyed/unkey/pkg/uid"
 	"github.com/unkeyed/unkey/svc/ctrl/integration/seed"
