+++ svc/ctrl/integration/harness/BUILD.bazel
@@ -13,9 +13,9 @@ go_library(
         "//pkg/clickhouse",
         "//pkg/clickhouse/schema",
         "//pkg/clock",
+        "//pkg/db/sqlcomment",
         "//pkg/healthcheck",
         "//pkg/restate/admin",
-        "//pkg/sqlcomment",
         "//pkg/testutil/containers",
         "//svc/ctrl/integration/seed",
         "//svc/ctrl/internal/db",
