+++ svc/ctrl/worker/BUILD.bazel
@@ -22,6 +22,7 @@ go_library(
         "//pkg/clickhouse/schema",
         "//pkg/clock",
         "//pkg/config",
+        "//pkg/db/sqlcomment",
         "//pkg/healthcheck",
         "//pkg/logger",
         "//pkg/otel",
@@ -31,7 +32,6 @@ go_library(
         "//pkg/restate/admin",
         "//pkg/rpc/interceptor",
         "//pkg/runner",
-        "//pkg/sqlcomment",
         "//pkg/uid",
         "//svc/ctrl/internal/auditlogs",
         "//svc/ctrl/internal/db",
