+++ svc/frontline/BUILD.bazel
@@ -27,6 +27,7 @@ go_library(
         "//pkg/config",
         "//pkg/counter",
         "//pkg/db",
+        "//pkg/db/sqlcomment",
         "//pkg/logger",
         "//pkg/otel",
         "//pkg/pprof",
@@ -36,7 +37,6 @@ go_library(
         "//pkg/rbac",
         "//pkg/rpc/interceptor",
         "//pkg/runner",
-        "//pkg/sqlcomment",
         "//pkg/tls",
         "//pkg/uid",
         "//pkg/zen",
